+++ nhais-adaptor/adaptor/outgoing/interchange_adaptor.py
@@ -7,6 +7,7 @@ from edifact.outgoing.models.interchange import Interchange
 from edifact.outgoing.models.message import Messages
 from adaptor.fhir_helpers.fhir_creators import ParameterName
 import adaptor.outgoing.common.common_adaptor as common_adaptor
+import adaptor.outgoing.common.date_formatter as date_formatter
 
 
 class InterchangeAdaptor:
@@ -26,6 +27,7 @@ class InterchangeAdaptor:
         sender_cypher = finders.get_parameter_value(fhir_operation, parameter_name=ParameterName.SENDER_CYPHER)
         nhais_cypher = finders.get_parameter_value(fhir_operation, parameter_name=ParameterName.NHAIS_CYPHER)
         recipient = common_adaptor.generate_recipient_from(nhais_cypher)
+        formatted_date_time = date_formatter.format_date(date_time=fhir_operation.date.as_json())
 
         message_adaptor = self.operation_dict[fhir_operation.name]["messageAdaptor"](fhir_operation)
 
@@ -33,7 +35,7 @@ class InterchangeAdaptor:
 
         interchange = Interchange(sender=sender_cypher, recipient=recipient,
                                   sequence_number=interchange_sequence_number,
-                                  date_time=fhir_operation.date.as_json(), messages=messages)
+                                  date_time=formatted_date_time, messages=messages)
 
         edifact_interchange = (sender_cypher, recipient, interchange_sequence_number, interchange.to_edifact())
