+++ nhais-adaptor/edifact/outgoing/models/interchange.py
@@ -1,4 +1,4 @@
-import edifact.helpers.date_formatter as date_formatter
+import adaptor.outgoing.common.date_formatter as date_formatter
 from edifact.outgoing.models.segment import Segment, SegmentCollection
 
 
@@ -18,8 +18,7 @@ class InterchangeHeader(Segment):
         :param date_time: the date time stamp of the interchange header
         :param sequence_number: a unique reference of the interchange
         """
-        formatted_date_time = date_formatter.format_date(date_time=date_time)
-        segment_value = f"UNOA:2+{sender}+{recipient}+{formatted_date_time}+{sequence_number}++FHSREG"
+        segment_value = f"UNOA:2+{sender}+{recipient}+{date_time}+{sequence_number}++FHSREG"
         super().__init__(key=self.SEGMENT_KEY, value=segment_value)
